Improved node-level cache. Cache ids are now based on a unique version number derived from the index. Added a service to track changes in an index and compute a version. Closes #73.

+++ src/main/java/solutions/siren/join/index/query/FieldDataTermsQuery.java
@@ -63,7 +63,7 @@ public abstract class FieldDataTermsQuery extends Query implements Accountable {
   /**
    * The cache key for this query
    */
-  protected final int cacheKey;
+  protected final long cacheKey;
 
   private static final ESLogger logger = Loggers.getLogger(FieldDataTermsQuery.class);
 
@@ -76,7 +76,7 @@ public abstract class FieldDataTermsQuery extends Query implements Accountable {
    * @param cacheKey      A unique key to use for caching this query.
    * @return the query.
    */
-  public static FieldDataTermsQuery newLongs(final byte[] encodedTerms, final IndexNumericFieldData fieldData, final int cacheKey) {
+  public static FieldDataTermsQuery newLongs(final byte[] encodedTerms, final IndexNumericFieldData fieldData, final long cacheKey) {
     return new LongsFieldDataTermsQuery(encodedTerms, fieldData, cacheKey);
   }
 
@@ -89,14 +89,14 @@ public abstract class FieldDataTermsQuery extends Query implements Accountable {
    * @param cacheKey      A unique key to use for caching this query.
    * @return the query.
    */
-  public static FieldDataTermsQuery newBytes(final byte[] encodedTerms, final IndexFieldData fieldData, final int cacheKey) {
+  public static FieldDataTermsQuery newBytes(final byte[] encodedTerms, final IndexFieldData fieldData, final long cacheKey) {
     return new BytesFieldDataTermsQuery(encodedTerms, fieldData, cacheKey);
   }
 
   /**
    * Creates a new {@link FieldDataTermsQuery} from the given field data.
    */
-  public FieldDataTermsQuery(final byte[] encodedTerms, final IndexFieldData fieldData, final int cacheKey) {
+  public FieldDataTermsQuery(final byte[] encodedTerms, final IndexFieldData fieldData, final long cacheKey) {
     this.encodedTerms = encodedTerms;
     this.fieldData = fieldData;
     this.cacheKey = cacheKey;
@@ -119,7 +119,7 @@ public abstract class FieldDataTermsQuery extends Query implements Accountable {
   @Override
   public int hashCode() {
     int hashcode = super.hashCode();
-    hashcode = 31 * hashcode + cacheKey; // relies on the cache key instead of the encodedTerms for hashcode
+    hashcode = 31 * hashcode + ((int) cacheKey); // relies on the cache key instead of the encodedTerms for hashcode
     return hashcode;
   }
 
@@ -193,7 +193,7 @@ public abstract class FieldDataTermsQuery extends Query implements Accountable {
      *
      * @param fieldData
      */
-    public LongsFieldDataTermsQuery(final byte[] encodedTerms, final IndexFieldData fieldData, final int cacheKey) {
+    public LongsFieldDataTermsQuery(final byte[] encodedTerms, final IndexFieldData fieldData, final long cacheKey) {
       super(encodedTerms, fieldData, cacheKey);
     }
 
@@ -260,7 +260,7 @@ public abstract class FieldDataTermsQuery extends Query implements Accountable {
      *
      * @param fieldData
      */
-    public BytesFieldDataTermsQuery(final byte[] encodedTerms, final IndexFieldData fieldData, final int cacheKey) {
+    public BytesFieldDataTermsQuery(final byte[] encodedTerms, final IndexFieldData fieldData, final long cacheKey) {
       super(encodedTerms, fieldData, cacheKey);
     }
 
@@ -325,9 +325,9 @@ public abstract class FieldDataTermsQuery extends Query implements Accountable {
    */
   private static class CacheKeyFieldDataTermsQuery extends MultiTermQuery {
 
-    private final int cacheKey;
+    private final long cacheKey;
 
-    public CacheKeyFieldDataTermsQuery(int cacheKey) {
+    public CacheKeyFieldDataTermsQuery(long cacheKey) {
       super("");
       this.cacheKey = cacheKey;
     }
@@ -354,7 +354,7 @@ public abstract class FieldDataTermsQuery extends Query implements Accountable {
     public int hashCode() {
       final int prime = 31;
       int result = 1;
-      result = prime * result + cacheKey;
+      result = prime * result + ((int) cacheKey);
       return result;
     }
