+++ src/main/java/solutions/siren/join/index/query/FieldDataTermsQueryParser.java
@@ -62,7 +62,7 @@ public class FieldDataTermsQueryParser implements QueryParser {
 
     String queryName = null;
     byte[] value = null;
-    Integer cacheKey = null;
+    Long cacheKey = null;
 
     token = parser.nextToken();
     if (token == XContentParser.Token.START_OBJECT) {
@@ -76,7 +76,7 @@ public class FieldDataTermsQueryParser implements QueryParser {
           } else if ("_name".equals(currentFieldName)) {
               queryName = parser.text();
           } else if ("_cache_key".equals(currentFieldName) || "_cacheKey".equals(currentFieldName)) {
-              cacheKey = parser.intValue();
+              cacheKey = parser.longValue();
           } else {
             throw new QueryParsingException(parseContext, "[fielddata_terms] filter does not support [" + currentFieldName + "]");
           }
@@ -113,7 +113,7 @@ public class FieldDataTermsQueryParser implements QueryParser {
   }
 
   private final Query toFieldDataTermsQuery(MappedFieldType fieldType, IndexFieldData fieldData,
-                                            byte[] encodedTerms, int cacheKey) {
+                                            byte[] encodedTerms, long cacheKey) {
     Query query = null;
 
     if (fieldType instanceof NumberFieldMapper.NumberFieldType) {
